pyproject.toml
[tool.poetry]
name = "app"
version = "1.0.0"
description = ""
authors = ["Oliver Arthur <contact@oliverarthurnardi.com>"]

[tool.poetry.dependencies]
python = "3.10"
fastapi = "0.70.0"
uvicorn = "0.15.0"
alembic = "1.7.4"
psycopg2 = "2.9.1"
SQLAlchemy = "1.4.26"
email-validator = "1.1.3"
starlette = "0.16.0"
urllib3 = "1.26.7"
pydantic = "1.8.2"
passlib = {version = "1.7.4", optional = true}
bcrypt = {version = "3.2.0", optional = true}
python-jose = {version = "3.3.0", optional = true}
cryptography = {version = "35.0.0", optional = true}
python-multipart = "0.0.5"
python-dotenv = "0.19.1"
toolz = "0.11.2"

[tool.poetry.dev-dependencies]
black = {version = "21.10b0", allow-prereleases = false}
flake8 = {version = "3.7.7", allow-prereleases = false}
mypy = {version = "0.761", allow-prereleases = false}
isort = {version = "5.10.1", allow-prereleases = false}
lxml = {version = "4.6.4", allow-prereleases = false}

[tool.poetry.scripts]
run-server = "scripts:server"
run-black = "scripts:black"
run-isort = "scripts:isort"
run-flake8 = "scripts:flake8"
run-mypy = "scripts:mypy"

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"


